+++ apps/bbsync/exceptions.py
@@ -7,3 +7,10 @@ class SRTNotesValidationError(Exception):
     """
     SRT notes JSON valition error exception class
     """
+
+
+class ProductDataError(Exception):
+    """
+    product data error exception class
+    for error in either the data or the product definitions
+    """