+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement validation for services related products with WONTREPORT resolution (OSIDB-362)
 - Implement validation for combinations of affectedness and resolution (OSIDB-360)
 - Implement a new API for getting a list of all supported products (PSINSIGHTS-593)
+- Implement CC list builder in Bugzilla backwards sync (OSIDB-386)
 
 ### Changed
 - Change logging of celery and django to filesystem (OSIDB-418)